+++ lib/src/generators.dart
@@ -2,11 +2,8 @@
 
 import 'dart:convert';
 
-import 'package:meta/meta.dart';
 import 'package:path/path.dart' as _path;
 
-part 'generators/stage_builder.dart';
 part 'generators/comments_generator.dart';
 part 'generators/generator.dart';
-part 'generators/items_generator.dart';
 part 'generators/part_generator.dart';