+++ lib/src/json_generator/json_object_generator.dart
@@ -3,24 +3,31 @@
 part of '../json_generator.dart';
 
 class JsonObjectGenerator extends Generator<Class> with CommentsGenerator {
+  final bool checkNullSafety;
+
   final bool immutable;
+
+  final int index;
 
   final JsonObject jsonObject;
 
-  JsonObjectGenerator({@required this.immutable, @required this.jsonObject});
+  JsonObjectGenerator(
+      {@required this.checkNullSafety,
+      @required this.index,
+      @required this.immutable,
+      @required this.jsonObject});
 
   @override
-  Class generate(StageBuilder builder) {
-    return builder.build(
-        'JSON object \'${jsonObject.name}\'', () => _build(builder));
-  }
-
-  bool isImmutable() {
-    if (jsonObject.immutable != null) {
-      return jsonObject.immutable;
-    }
-
-    return immutable == true;
+  Class generate() {
+    return Class((cb) {
+      _addComments(cb);
+      _addAnnotations(cb);
+      _setName(cb);
+      _addFields(cb);
+      _addContructor(cb);
+      _addFactoryFromJson(cb);
+      _addMethodToJson(cb);
+    });
   }
 
   void _addAnnotation(ClassBuilder cb, Expression annotation) {
@@ -46,7 +53,7 @@ class JsonObjectGenerator extends Generator<Class> with CommentsGenerator {
 
   void _addContructor(ClassBuilder cb) {
     final result = Constructor((b) {
-      final properties = jsonObject.properties ?? [];
+      final properties = jsonObject.properties;
       for (final property in properties) {
         b.optionalParameters.add(Parameter((b) {
           b.named = true;
@@ -81,17 +88,19 @@ class JsonObjectGenerator extends Generator<Class> with CommentsGenerator {
     cb.constructors.add(result);
   }
 
-  void _addFields(StageBuilder b, ClassBuilder cb) {
-    final result = ItemsGenerator(
-        stage: 'Property',
-        input: jsonObject.properties,
-        build: (Property property) {
-          final generator =
-              PropertyGenerator(immutable: isImmutable(), property: property);
-          return generator.generate(b);
-        }).generate(b);
-
-    cb.fields.addAll(result);
+  void _addFields(ClassBuilder cb) {
+    final list = jsonObject.properties;
+    for (var i = 0; i < list.length; i++) {
+      final element = list[i];
+      final g = PropertyGenerator(
+          immutable: _isImmutable(),
+          index: i,
+          objectName: jsonObject.name,
+          property: element,
+          checkNullSafety: _needCheckNullSafety());
+      final result = g.generate();
+      cb.fields.add(result);
+    }
   }
 
   void _addMethodToJson(ClassBuilder cb) {
@@ -107,24 +116,24 @@ class JsonObjectGenerator extends Generator<Class> with CommentsGenerator {
     cb.methods.add(result);
   }
 
-  Class _build(StageBuilder b) {
-    return Class((cb) {
-      _addComments(cb);
-      _addAnnotations(cb);
-      _setName(cb);
-      _addFields(b, cb);
-      _addContructor(cb);
-      _addFactoryFromJson(cb);
-      _addMethodToJson(cb);
-    });
+  List<String> _getAnnotations() {
+    return jsonObject.annotations;
   }
 
-  List<String> _getAnnotations() {
-    return jsonObject.annotations ?? [];
+  bool _isImmutable() {
+    if (jsonObject.immutable != null) {
+      return jsonObject.immutable;
+    }
+
+    return immutable == true;
+  }
+
+  bool _needCheckNullSafety() {
+    return checkNullSafety == true;
   }
 
   void _setName(ClassBuilder b) {
-    final name = getField(jsonObject.name, 'name');
+    final name = getField(jsonObject.name, 'The name of JSON object with index $index is not specified');
     b.name = name;
   }
 }