+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.2
+
+- Reduced and improved the source code of the JSON generator
+- Added the ability to specify the need to check JSON objects for null safety when generating code to the JSON generator specification (property `checkNullSafety` of object `JsonObjects`).
+
 ## 0.2.1
 
 - Fixed bug with `Uri` schema `file` when working with `.packages` file.